solvedMess

+++ src/test/java/com/cybertek/tests/day11_file_upload/FileUploadTest.java
@@ -17,7 +17,7 @@ public class FileUploadTest {
 
     @AfterMethod
     public void tearDown() {
-        Sleeper.sleep(0.5);
+        Sleeper.sleep((int) 0.5);
         driver.quit();
     }
